backend/pyproject.toml
[project]
name = "speak-to-llm-backend"
version = "0.1.0"
description = "Voice-to-LLM backend: Whisper transcription + LM Studio proxy"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "openai-whisper>=20231117",
    "numpy>=1.26.0",
    "openai>=1.30.0",
    "pydantic-settings>=2.2.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "httpx>=0.27.0",
    "pytest-asyncio>=0.23.0",
]


[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
